make sure Context sets required fields
While working on getting the [Go HTTP Routing Benchmark](https://github.com/julienschmidt/go-http-routing-benchmark)
to compile / run, an issue with the `bear` package was being observed.
Specifically, both the `http.ResponseWriter` and `*http.Request` sent
to the HandleFunc were set to `nil`. Upon digging in to the codebase
it was found that these two `Context{}` structs were being created
incorrectly.

It may be worth mentioning that I am not a user of `bear`, so I wouldn't
be able to determine whether these are the right changes to be making.
This should probably also have a regression test written for it.

Please let me know if any thing else is needed to merge this changeset.